introduced taskboard.board_loader to get a board
It's imperfect - it's a service locator, but at this point
it feels too early to any kind of full DI solution - plus
django in general doesn't tend to do that (it would use
settins for it, but we can get to that later if needed).

For now, it's improved - the board builder in the test is
responsible for building it up, so thus it should be responsible
for getting it injected to the appropriate places - the tests
are cleaner.

Next up: remove the board argument passed to all the other helpers